requirement.txt
asgiref==3.4.0
boto3==1.17.102
botocore==1.20.102
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
Django==3.2.4
django-cors-headers==3.7.0
django-datefilterspec==1.3.0
django-daterange-filter==1.3.0
django-extensions==3.1.3
django-filter==2.4.0
django-redis==5.0.0
django-redis-cache==3.0.0
django-ticketing==0.7.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-proxy-pagination==0.2.0
fcm-django==0.3.10
firebase-admin==5.0.1
google-api-core==1.30.0
google-api-python-client==2.11.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.1
google-cloud-firestore==2.1.3
google-cloud-storage==1.39.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.38.1
httplib2==0.19.1
idna==2.10
jmespath==0.10.0
msgpack==1.0.2
osticket==0.2.1
packaging==20.9
Pillow==8.2.0
proto-plus==1.19.0
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.4.9
PyJWT==2.1.0
pyosticket==0.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
rest-framework-pagination==0.0.1
rsa==4.7.2
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
